+++ app/src/main/java/com/nsd/talk/ui/chat/ChatViewModel.kt
@@ -1,4 +1,4 @@
-package com.nsd.talk.ui
+package com.nsd.talk.ui.chat
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
