+++ app/src/main/java/com/nsd/talk/ui/chatroom/ChatRoomFragment.kt
@@ -1,17 +1,13 @@
-package com.nsd.talk.ui
+package com.nsd.talk.ui.chatroom
 
 import android.content.Context
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.viewModels
 import androidx.lifecycle.LifecycleObserver
-import androidx.lifecycle.Observer
-import androidx.recyclerview.widget.LinearLayoutManager
 import com.nsd.talk.R
 import com.nsd.talk.databinding.FragmentChatRoomBinding
