+++ app/src/main/java/com/nsd/talk/ui/friend/FriendViewModel.kt
@@ -1,4 +1,4 @@
-package com.nsd.talk.ui
+package com.nsd.talk.ui.friend
 
 import androidx.lifecycle.ViewModel
