+++ app/src/main/res/layout/activity_chat.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.ChatRoomFragment">
+    tools:context=".ui.chatroom.ChatRoomFragment">
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
