+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context=".ui.MainActivity">
+    tools:context=".ui.main.MainActivity">
 
 
     <FrameLayout
