+++ app/src/main/res/layout/fragment_friend.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.FriendFragment">
+    tools:context=".ui.friend.FriendFragment">
 
     <TextView
         android:layout_width="match_parent"
